+++ src/Ankimon/__init__.py
@@ -139,6 +139,7 @@ try:
     from .functions.badges_functions import check_badges, check_for_badge, receive_badge, handle_achievements, check_and_award_badges
     from .functions.pokemon_functions import get_random_moves_for_pokemon
     from .functions.encounter_functions import choose_random_pkmn_from_tier
+    from .functions.pokemon_showdown_functions import export_to_pkmn_showdown, export_all_pkmn_showdown
     from .functions.pokemon_functions import (
         pick_random_gender,
         find_experience_for_level,
@@ -1433,181 +1434,6 @@ def on_review_card(*args):
 gui_hooks.reviewer_did_answer_card.append(on_review_card)
 
 life_bar_injected = False
-
-def export_to_pkmn_showdown():
-    # Create a main window
-    window = QDialog(mw)
-    window.setWindowTitle("Export Pokemon to Pkmn Showdown")
-    for stat, value in main_pokemon.ev.items():
-        if value == 0:
-            main_pokemon.ev[stat] += 1
-    # Format the Pokemon info
-    pokemon_info = "{} ({})\nAbility: {}\nLevel: {}\nType: {}\nEVs: {} HP / {} Atk / {} Def / {} SpA / {} SpD / {} Spe\n IVs: {} HP / {} Atk / {} Def / {} SpA / {} SpD / {} Spe ".format(
-        main_pokemon.name,
-        main_pokemon.gender,
-        main_pokemon.ability,
-        main_pokemon.level,
-        main_pokemon.type,
-        main_pokemon.ev["hp"],
-        main_pokemon.ev["atk"],
-        main_pokemon.ev["def"],
-        main_pokemon.ev["spa"],
-        main_pokemon.ev["spd"],
-        main_pokemon.ev["spe"],
-        main_pokemon.iv["hp"],
-        main_pokemon.iv["atk"],
-        main_pokemon.iv["def"],
-        main_pokemon.iv["spa"],
-        main_pokemon.iv["spd"],
-        main_pokemon.iv["spe"]
-    )
-    for attack in main_pokemon.attacks:
-        pokemon_info += f"\n- {attack}"
-    # Information label
-    info = "Pokemon Infos have been Copied to your Clipboard! \nNow simply paste this text into Teambuilder in PokemonShowdown. \nNote: Fight in the [Gen 9] Anything Goes - Battle Mode"
-    info += f"\n Your Pokemon is considered Tier: {search_pokedex(main_pokemon.name.lower(), 'tier')} in PokemonShowdown"
-    # Create labels to display the text
-    label = QLabel(pokemon_info)
-    info_label = QLabel(info)
-
-    # Align labels
-    label.setAlignment(Qt.AlignmentFlag.AlignCenter)  # Align center
-    info_label.setAlignment(Qt.AlignmentFlag.AlignCenter)  # Align center
-
-    # Create a layout and add the labels
-    layout = QVBoxLayout()
-    layout.addWidget(info_label)
-    layout.addWidget(label)
-
-    # Set the layout for the main window
-    window.setLayout(layout)
-
-    # Copy text to clipboard in Anki
-    mw.app.clipboard().setText(pokemon_info)
-
-    # Show the window
-    window.show()
-
-def save_error_code(error_code):
-    error_fix_msg = ""
-    try:
-        # Find the position of the phrase "can't be transferred from Gen"
-        index = error_code.find("can't be transferred from Gen")
-
-        # Extract the substring starting from this position
-        relevant_text = error_code[index:]
-
-        # Find the first number in the extracted text (assuming it's the generation number)
-        generation_number = int(''.join(filter(str.isdigit, relevant_text)))
-
-        # Show the generation number
-        error_fix_msg += (f"\n Please use Gen {str(generation_number)[0]} or lower")
-
-        index = error_code.find("can't be transferred from Gen")
-
-        # Extract the substring starting from this position
-        relevant_text = error_code[index:]
-
-        # Find the first number in the extracted text (assuming it's the generation number)
-        generation_number = int(''.join(filter(str.isdigit, relevant_text)))
-
-        error_fix_msg += (f"\n Please use Gen {str(generation_number)[0]} or lower")
-
-    except Exception as e:
-        logger.log_and_showinfo("info",f"An error occurred: {e}")
-
-    logger.log_and_showinfo("info",f"{error_fix_msg}")
-
-def export_all_pkmn_showdown():
-    # Create a main window
-    export_window = QDialog()
-    #export_window.setWindowTitle("Export Pokemon to Pkmn Showdown")
-
-    # Information label
-    info = "Pokemon Infos have been Copied to your Clipboard! \nNow simply paste this text into Teambuilder in PokemonShowdown. \nNote: Fight in the [Gen 7] Anything Goes - Battle Mode"
-    info_label = QLabel(info)
-
-    # Get all pokemon data
-    pokemon_info_complete_text = ""
-    try:
-        with (open(mypokemon_path, "r", encoding="utf-8") as json_file):
-            captured_pokemon_data = json.load(json_file)
-
-            # Check if there are any captured Pokémon
-            if captured_pokemon_data:
-                # Counter for tracking the column position
-                column = 0
-                row = 0
-                for pokemon in captured_pokemon_data:
-                    pokemon_name = pokemon['name']
-                    pokemon_level = pokemon['level']
-                    pokemon_ability = pokemon['ability']
-                    pokemon_type = pokemon['type']
-                    pokemon_type_text = pokemon_type[0].capitalize()
-                    if len(pokemon_type) > 1:
-                        pokemon_type_text = ""
-                        pokemon_type_text += f"{pokemon_type[0].capitalize()}"
-                        pokemon_type_text += f" {pokemon_type[1].capitalize()}"
-                    pokemon_stats = pokemon['stats']
-                    pokemon_hp = pokemon_stats["hp"]
-                    pokemon_attacks = pokemon['attacks']
-                    pokemon_ev = pokemon['ev']
-                    pokemon_iv = pokemon['iv']
-
-                    pokemon_info = "{} \nAbility: {}\nLevel: {}\nType: {}\nEVs: {} HP / {} Atk / {} Def / {} SpA / {} SpD / {} Spe\n IVs: {} HP / {} Atk / {} Def / {} SpA / {} SpD / {} Spe \n".format(
-                        pokemon_name,
-                        pokemon_ability.capitalize(),
-                        pokemon_level,
-                        pokemon_type_text,
-                        pokemon_stats["hp"],
-                        pokemon_stats["atk"],
-                        pokemon_stats["def"],
-                        pokemon_stats["spa"],
-                        pokemon_stats["spd"],
-                        pokemon_stats["spe"],
-                        pokemon_iv["hp"],
-                        pokemon_iv["atk"],
-                        pokemon_iv["def"],
-                        pokemon_iv["spa"],
-                        pokemon_iv["spd"],
-                        pokemon_iv["spe"]
-                    )
-                    for attack in pokemon_attacks:
-                        pokemon_info += f"- {attack}\n"
-                    pokemon_info += "\n"
-                    pokemon_info_complete_text += pokemon_info
-
-                    # Create labels to display the text
-                    #label = QLabel(pokemon_info_complete_text)
-                    # Align labels
-                    #label.setAlignment(Qt.AlignmentFlag.AlignCenter)  # Align center
-                    info_label.setAlignment(Qt.AlignmentFlag.AlignCenter)  # Align center
-
-                    # Create an input field for error code
-                    error_code_input = QLineEdit()
-                    error_code_input.setPlaceholderText("Enter Error Code")
-
-                    # Create a button to save the input
-                    save_button = QPushButton("Fix Pokemon Export Code")
-
-                    # Create a layout and add the labels, input field, and button
-                    layout = QVBoxLayout()
-                    layout.addWidget(info_label)
-                    #layout.addWidget(label)
-                    layout.addWidget(error_code_input)
-                    layout.addWidget(save_button)
-
-                    # Copy text to clipboard in Anki
-                    mw.app.clipboard().setText(pokemon_info_complete_text)
-
-        save_button.clicked.connect(lambda: save_error_code(error_code_input.text()))
-
-        # Set the layout for the main window
-        export_window.setLayout(layout)
-
-        export_window.exec()
-    except Exception as e:
-        logger.log_and_showinfo("info",f"An error occurred: {e}")
 
 def flex_pokemon_collection():
     # Create a main window
